% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My First Functions}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a list with components 'x' and 'y', where 'y' is the square of 'x'
}
\description{
My First Functions
}
\examples{
myfirstfunction(1:10)
}
